# specs/001-bloom-tutor-app/contracts/api.yaml
openapi: 3.0.3
info:
  title: Bloom GCSE Mathematics Tutor API
  description: |
    FastAPI backend for Bloom tutoring application.
    Uses htmx for frontend interactions - most endpoints return HTML fragments.
  version: 1.0.0
  contact:
    name: Bloom Development Team
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /admin/syllabus/upload:
    post:
      tags: [admin]
      summary: Load syllabus from JSON file
      description: |
        Validates JSON structure (FR-019) and loads into database.
        Replaces existing topics/subtopics.
        Preserves student progress data.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: JSON file containing syllabus
      responses:
        '200':
          description: Syllabus loaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Loaded 3 topics, 12 subtopics"
        '400':
          description: Validation error with specific details (FR-019)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /admin/syllabus/validate:
    post:
      tags: [admin]
      summary: Validate syllabus JSON without loading
      description: Checks structure before committing to database
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Validation passed
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
                  topics_count:
                    type: integer
                    example: 3
                  subtopics_count:
                    type: integer
                    example: 12
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...
components:
  schemas:
...
    ValidationError:
      type: object
      properties:
        valid:
          type: boolean
          example: false
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "topics[2].subtopics[0].id"
              message:
                type: string
                example: "field required"
        details:
          type: string
          example: "Missing 'id' field in Topic 3, Subtopic 1"

# =============================================================================
# SECURITY (for future)
# =============================================================================

# Note: Current demo has no authentication.
# For production, add:
# securitySchemes:
#   bearerAuth:
#     type: http
#     scheme: bearer
# security:
#   - bearerAuth: []